requirements.txt
pydub
openai
python-dotenv
sentsplit
tiktoken
celery
Flask
Werkzeug
gunicorn